Passing recursive fibonacci
Needed to save $16 on the stack so that nested function calls
restore $16 as they roll back! $16 is used to preserve the value of $3,
which is used in the binary operators to store the value of the left-
hand side. When stacking + or - this isn't a problem (?) but for
functions it is